+++ Jumphosts/roles/ansible-role-ssh-add-jumphosts/tasks/jumphostlogin.sh
@@ -4,9 +4,11 @@ echo SSH_AUTH_SOCK=$SSH_AUTH_SOCK
 echo SSH_AGENT_PID=$SSH_AGENT_PID
 if [ ! -z "$endpoint_ssh_private_key_passphrase" ]; then
   echo endpoint=${endpoint_ssh_private_key##*/}
-  ssh-add ${endpoint_ssh_private_key}
-  rm -f ${endpoint_ssh_private_key}
-  echo Removed ${endpoint_ssh_private_key}
+  if [ -e "${endpoint_ssh_private_key}" ]; then
+    ssh-add ${endpoint_ssh_private_key}
+    rm -f ${endpoint_ssh_private_key}
+    echo Removed ${endpoint_ssh_private_key}
+  fi
 fi
 if [ ! -z "$jh1_ssh_private_key_passphrase" ]; then
   echo jh1=${jh1_ssh_private_key##*/}
